+++ node/InputNode/node_video.py
@@ -115,7 +115,7 @@ class FactoryNode:
         small_window_h = node._opencv_setting_dict["input_window_height"]
         use_pref_counter = node._opencv_setting_dict["use_pref_counter"]
 
-        black_image = np.zeros((small_window_h, small_window_w, 3))
+        black_image = np.zeros((small_window_h, small_window_w, 3), dtype=np.uint8)
         black_texture = node.convert_cv_to_dpg(
             black_image,
             small_window_w,
