+++ node/OverlayNode/node_puttext.py
@@ -89,7 +89,7 @@ class FactoryNode:
         small_window_h = node._opencv_setting_dict['process_height']
 
 
-        black_image = np.zeros((small_window_w, small_window_h, 3))
+        black_image = np.zeros((small_window_w, small_window_h, 3), dtype=np.uint8)
         black_texture = node.convert_cv_to_dpg(
             black_image,
             small_window_w,
