+++ node/VideoNode/node_dynamic_play.py
@@ -57,7 +57,7 @@ class FactoryNode:
         small_window_h = node._opencv_setting_dict['process_height']
 
         # Create black texture (height, width, channels)
-        black_image = np.zeros((small_window_h, small_window_w, 3))
+        black_image = np.zeros((small_window_h, small_window_w, 3), dtype=np.uint8)
         black_texture = node.convert_cv_to_dpg(
             black_image,
             small_window_w,
